Fixed the dropEvent when playlist is showed, and new slider class
There should be no more crashes when dropping music to the player while
playlist is open.
A new Slider class has been created for easy of use. It works exactly
like the previous QSliders, but the user is no longer restricted to only
slide the handle to move through; a click is enough. :)